+++ server/src/parser/ast/segments.ts
@@ -1,0 +1,9 @@
+import { AstNode } from "./nodes";
+
+export type SegmentsByName = Map<string, SegmentObject>
+
+export type SegmentObject = {
+    name: string,
+    definitions: AstNode[],
+    usages: AstNode[],
+}