+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## [Unreleased]
+
+### Added 
+
+- Work with segment names: autocompletion, usage, renaming
+
 ## [2.0.0] - 2024-01-21
 ### Braking change